requirements.txt
opencv-python~=4.6.0.66
opencv-contrib-python~=4.6.0.66
numpy~=1.23.4
Flask~=2.2.2
waitress~=2.1.2
Pillow~=9.3.0
socketio~=0.2.1